+++ common/src/main/java/org/onehippo/forge/ipfilter/common/IpFilterConfigLoader.java
@@ -28,13 +28,9 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentSkipListSet;
 
-import javax.jcr.Credentials;
-import javax.jcr.Node;
-import javax.jcr.NodeIterator;
-import javax.jcr.Repository;
-import javax.jcr.RepositoryException;
-import javax.jcr.Session;
+import javax.jcr.*;
 import javax.jcr.observation.Event;
 
 import org.apache.commons.lang.ArrayUtils;
@@ -72,6 +68,7 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
 
     private final Map<String, AuthObject> data = new ConcurrentHashMap<>();
 
+    private final Set<String> forwardedForHostHeaders = Collections.synchronizedSet(new HashSet<>());
     public IpFilterConfigLoader() {
 
         try {
@@ -217,7 +214,7 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
         } catch (IOException e) {
             log.error("Error loading {}", IpFilterConstants.PROPERTIES_NAME, e);
         }
-        
+
         return ImmutableListMultimap.of();
     }
 
@@ -287,6 +284,19 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
     private void parseConfig(final Node node) throws RepositoryException {
         final Map<String, AuthObject> objects = new HashMap<>();
         final Multimap<String, String> globalSettings = loadGlobalSettings();
+        forwardedForHostHeaders.clear();
+        if(node.hasProperty(IpFilterConstants.CONFIG_FORWARDED_HOST_HEADER)){
+            final Value[] property = node.getProperty(IpFilterConstants.CONFIG_FORWARDED_HOST_HEADER).getValues();
+            for (Value value : property) {
+                final String v = value.getString();
+                if (Strings.isNullOrEmpty(v)) {
+                    log.debug("Adding  CONFIG_FORWARDED_FOR_HEADERS: {}", v);
+                    forwardedForHostHeaders.add(v);
+                }
+            }
+        }
+        // always add default:
+        forwardedForHostHeaders.add(IpFilterConstants.HEADER_X_FORWARDED_HOST);
         final NodeIterator nodes = node.getNodes();
         while (nodes.hasNext()) {
             final Node configNode = nodes.nextNode();
@@ -297,6 +307,7 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
         }
         data.clear();
         data.putAll(objects);
+        // populate
     }
 
     private AuthObject parse(final Node node, final Multimap<String, String> globalSettings) throws RepositoryException {
@@ -339,8 +350,9 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
 
         final String forwardHeader = JcrUtils.getStringProperty(node, IpFilterConstants.CONFIG_FORWARDED_FOR_HEADER, IpFilterConstants.HEADER_X_FORWARDED_FOR);
         final boolean matchAll = JcrUtils.getBooleanProperty(node, IpFilterConstants.CONFIG_MATCH_ALL, false);
+        final boolean cacheEnabled = JcrUtils.getBooleanProperty(node, IpFilterConstants.CONFIG_CACHE_ENABLED, true);
 
-        return new AuthObject(ignoredPathSet, hostSet, rangesSet, ignoredHeaders, allowCmsUsers, forwardHeader, matchAll);
+        return new AuthObject(ignoredPathSet, hostSet, rangesSet, ignoredHeaders, allowCmsUsers, forwardHeader,cacheEnabled, matchAll);
     }
 
     private Map<String, Set<String>> parseHeaders(final Node root) throws RepositoryException {
@@ -401,5 +413,9 @@ public abstract class IpFilterConfigLoader implements FileChangeObserver {
 
     public void setCredentials(final Credentials credentials) {
         this.credentials = credentials;
+    }
+
+    public Set<String> getForwardedForHostHeaders() {
+        return forwardedForHostHeaders;
     }
 }