+++ common/src/main/java/org/onehippo/forge/ipfilter/common/IpFilterUtils.java
@@ -62,16 +62,7 @@ public final class IpFilterUtils {
         return request.getRemoteAddr();
     }
 
-    public static String getHost(final HttpServletRequest request) {
-        final String hostHeader = request.getHeader(IpFilterConstants.HEADER_X_FORWARDED_HOST);
-        if (Strings.isNullOrEmpty(hostHeader)) {
-            final String remoteHost = request.getRemoteHost();
-            log.debug("Missing header {}, using: {}", IpFilterConstants.HEADER_X_FORWARDED_HOST, remoteHost);
-            return remoteHost;
-        }
-        return hostHeader;
 
-    }
 
     public static String getPath(final HttpServletRequest request) {
         return request.getRequestURI().substring(request.getContextPath().length());